Extract out test_participant_emails.py
